pyproject.toml
[project]
name = "emby-collection-creator"
version = "0.1.0"
description = "MCP server for AI-powered Emby collection management"
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.4.0",
    "httpx>=0.28.1",
    "mcp>=1.22.0",
    "pillow>=11.0.0",
]

[project.scripts]
emby-mcp = "emby_collection_creator.mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/emby_collection_creator"]

[tool.hatch.build.targets.editable]
sources = ["src"]
